refactor(token-count): reuse cancellation and test fixtures
